REF: stats.unuran: Refactor the UNU.RAN wrapper to avoid memory leaks
This commit refactors the entire UNU.RAN wrapper to use `MessageStream`
to report errors instead of non-local returns. One of the downside of this
approach is that UNU.RAN uses a global thread-unsafe file streams which
makes error reporting thread unsafe. Hence, a module-level lock is used
to hold the file stream to report the errors to and achieve thread-safety.