REF, TST: stats: accept a single `dist` object instead of seperate method parameters.
This refactors the public API of the methods TransformedDensityRejection and DiscreteAliasUrn
method to accept a single `dist` object that contains all the required methods. This closely
follows the design of gh-13319. So, for example, previously one had to write required functions
seperately and pass them to the method for setup:

    >>> from scipy.stats import TransformedDensityRejection
    >>>
    >>> pdf = lambda x: 1-x*x
    >>> dpdf = lambda x: -2*x
    >>>
    >>> rng = TransformedDensityRejection(pdf, dpdf, domain=(-1,1), seed=123)

But, with the new API, one can just pass a `dist` object with all those methods bundled in it:

    >>> class MyDist:
    ...     def pdf(self, x):
    ...         return 1-x*x
    ...     def dpdf(self, x):
    ...         return -2*x
    ...
    >>> dist = MyDist()
    >>> rng = TransformedDensityRejection(dist, domain=(-1,1), seed=123)

Tests have been alterned to use the new API. New tests for the API itself need to be added.