REF, TST: evaluate the PMF in the domain to report helpful errors
UNU.RAN was unable to recognize inf, nan and all zeros in the PV and threw unhelpful
"unknown error". Hence, the DAU method has been refactored to evaluate the PMF during
argument validation and check the values in the PV to report helpful errors. This is
possible to do as DAU only works for distribution with finite domain.

The tests have also been changed to incorporate this refactor.

Docs of the DAU method have also been updated.