BUG: stats: fix failing tests
Two tests failed:
  1. TestDiscreteAliasUrn::test_basic
  2. Both `test_seed` tests

The former was because a non-string distribution name was present in `distdiscrete` which led
to problems with getattr. The straight-forward fix for this was to skip for non-string `distname`s

The latter was due to a more subtle reason. I used global variable to store the NumPy RNG for sampling
when `np.__version__ < 1.19`. Because of this, when a new NumPy RNG was set, all the methods would start
using that RNG and seeding would immediately break. This has been fixed by using a seperate function in the
base-class `Method` and a global function for default RNG.