TST: stats: add 'randint' to basic_fail_dists
'randint' was failing with an "unknown error" on 32-bit Ubuntu with NumPy 1.16.5.
The error is occurring on line 723 in file src/methods/dau.c. Looks like it might
be due to floating-point errors. As this is not inside SciPy (and also only exists
on a very specific platform and numpy version), I have skipped this test case.

Also, previously, a test was skipped if the distribution name wasn't string. This
has been corrected to run the test on that test case.